[UI] Phase 5.b.5: Migrate environment/connection modals to shared primitives (#18753)
Migrates three environment/connection modals onto the shared `Modal`
primitive from `@/components/shared/Modal` and relocates them into the
target domain folders:

- `shared/Modal/EnvironmentModal.tsx` -> `environments/EnvironmentFormModal.tsx`
- `shared/Modal/ConnectionModal.tsx` -> `connections/ConnectionFormModal.tsx`
- `connections/meshSync/RegisterConnectionModal.tsx` -> `connections/RegisterConnectionModal.tsx`

Each new file:
- Uses the shared `Modal` (consistent `isOpen`/`onClose`/`title`/`size`/
  `headerIcon`) instead of importing `Modal`/`ModalBody` directly from
  `@sistent/sistent`.
- Routes theme primitives (`useTheme`, `styled`) through `@/theme`.
- Replaces ad-hoc inline `style={{}}` with `styled('div')` wrappers.
- Preserves the legacy consumer-facing prop names so the three callers
  only need to update their import paths.

Consumer updates:
- `designs/lifecycle/DeployStepper.tsx`
- `layout/Header/Header.tsx`
- `connections/meshSync/index.tsx` (+ corresponding `index.test.tsx`
  mock path)

The `ConnectionTable` no longer accepts `meshsyncControllerState` /
`connectionMetadataState` props (it sources them from Redux directly).
`ConnectionFormModal` accepts those props for backwards compatibility
but no longer threads them through.

`eslint.config.js`: removed the two legacy allowlist entries
(`components/shared/Modal/ConnectionModal.tsx`,
`components/connections/meshSync/RegisterConnectionModal.tsx`); the
rewritten files are clean. `EnvironmentModal.tsx` was never on the list.

Closes #18753
Refs #18661 #18654
