+++ ui/components/connections/meshSync/index.tsx
@@ -20,7 +20,7 @@ import {
   getColumnValue,
   getVisibilityColums,
 } from '../../../utils/utils';
-import RegisterConnectionModal from './RegisterConnectionModal';
+import RegisterConnectionModal from '../RegisterConnectionModal';
 import { CONNECTION_STATES, MESHSYNC_STATES } from '../../../utils/Enum';
 import { getResponsiveColumnVisibility } from '../../../utils/responsive-column';
 import { useWindowDimensions } from '../../../utils/dimension';
