+++ ui/eslint.config.js
@@ -249,7 +249,6 @@ const legacyInlineStyleOffenders = [
   'components/general/ConnectClustersBtn.tsx',
   'components/general/CreateDesignBtn.tsx',
   'components/shared/ErrorBoundary/ErrorBoundary.tsx',
-  'components/shared/Modal/ConnectionModal.tsx',
   'components/shared/Modal/Information/LegacyInfoModal.tsx',
   'components/shared/Modal/LegacyRJSFModal.tsx',
   'components/general/TipsCarousel.tsx',
@@ -359,7 +358,6 @@ const legacyInlineStyleOffenders = [
   'components/connections/common/index.tsx',
   'components/connections/index.tsx',
   'components/connections/meshSync/MeshSyncEmptyState.tsx',
-  'components/connections/meshSync/RegisterConnectionModal.tsx',
   'components/connections/meshSync/Stepper/Notification.tsx',
   'components/connections/meshSync/Stepper/StepperContent.tsx',
   'components/connections/meshSync/Stepper/StepperContentWrapper.tsx',
